Move components
This will move components to the parent folder.
