+++ components/index.js
@@ -1,0 +1,11 @@
+import Carousel from './Carousel';
+import Content from './Content';
+import Day from './Day';
+import GalleryCard from './GalleryCard';
+import Header from './Header';
+import Location from './Location';
+import Photo from './Photo';
+import Tags from './Tags';
+import Title from './Title';
+
+export { Carousel, Content, Day, GalleryCard, Header, Location, Photo, Tags, Title };
