+++ screens/ForgotPassword.js
@@ -1,7 +1,7 @@
 import { useRef, useState } from 'react';
 import { StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
 
-import Header from '../components/gallery/Header';
+import { Header } from '../components';
 import { Color, getKorErrorMsg } from '../components/util';
 import { auth } from '../firebaseConfig';
